fix logic error of including attempted problems as solved problems
Earlier the code included all attempted problems with ? infront of them in the UI as solved problems(the ones with a tick mark infront of them). This is not an expected behaviour. So, changing the logic of isSolved() is required and now the function only considers those problems with a span tag having the attribute value set to ac as solved problems